ADD) 미팅 방 참여, 퇴장 로직 구현

See merge request s09-webmobile1-sub2/S09P12A703!29

+++ gaebap/src/main/java/site/doggyyummy/gaebap/domain/meeting/dto/ModifyMeetingRequestDTO.java
@@ -21,7 +21,7 @@ public class ModifyMeetingRequestDTO {
     private String password;
 
     @JsonProperty("max_participant")
-    private Long maxParticipant;
+    private int maxParticipant;
 
     private String title;
